use millisecond precision for time properties

+++ lib/mixpanel-ruby/tracker.rb
@@ -165,7 +165,7 @@ module Mixpanel
       properties = {
         'distinct_id' => distinct_id,
         'token' => @token,
-        'time' => Time.now.to_i,
+        'time' => Time.now.to_f,
         'mp_lib' => 'ruby',
         '$lib_version' => Mixpanel::VERSION,
       }.merge(properties)
